pyproject.toml
[tool.poetry]
name = "fasthtml-tractor"
version = "0.1.0a1"
description = ""
authors = ["Philip Nuzhnyi <philip.nuzhnyy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-fasthtml = "^0.6.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
